+++ CHANGELOG.md
@@ -10,8 +10,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Support for using bech32 encoded private keys in `SuiConfig.user_config`
+- Support using Sui GraphQL in devnet
 
 ### Fixed
+
+- Overlap on Validator and ValidatorSet data class
 
 ### Changed